requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
pandas
prophet
plotly
matplotlib
scikit-learn
python-dotenv
db-dtypes